jsb/requirements.txt
aliyun-python-sdk-core-v3==2.13.3
beautifulsoup4==4.8.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
django==2.2.11
gevent==1.4.0
greenlet==0.4.15
idna==2.7
jmespath==0.9.4
mysqlclient==1.4.1
pycparser==2.19
pytz==2018.9
redis==3.2.1
requests==2.19.1
soupsieve==1.9.2
urllib3==1.23
xlwt==1.3.0
djangorestframework==3.11.0
django-oauth-toolkit==1.3.0
django-guardian==2.1.0
django-apscheduler==0.5.2
mysql-connector-python==8.0.23
DingtalkChatbot==1.5.2
